+++ .babelrc
@@ -1,5 +1,5 @@
 {
-    "presets": ["react", "es2015-loose"],
+    "presets": ["react", "es2015-loose", "stage-2"],
     "plugins": [
         'transform-class-properties',
         'transform-es3-member-expression-literals',
