+++ packages/react-data-grid/src/rxjs/Reducers.js
@@ -1,0 +1,8 @@
+import Rx from 'rxjs';
+import {cellReducer$} from './modules/Cell';
+
+const reducer$ = Rx.Observable.merge(
+  cellReducer$.map(reducer => ['cell', reducer]),
+);
+
+export default reducer$;
